refactor: Change project folder architecture
