+++ pkg/s3-proxy/bucket/utils.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/dustin/go-humanize"
-	"github.com/oxyno-zeta/s3-proxy/pkg/s3client"
+	"github.com/oxyno-zeta/s3-proxy/pkg/s3-proxy/s3client"
 )
 
 func setHeadersFromObjectOutput(w http.ResponseWriter, obj *s3client.GetOutput) {
