+++ pkg/s3-proxy/config/config_integration_test.go
