+++ pkg/s3-proxy/config/config_test.go
