+++ pkg/s3-proxy/config/validation.go
