+++ pkg/s3-proxy/config/validation_test.go
