+++ pkg/s3-proxy/metrics/client.go
