+++ pkg/s3-proxy/metrics/doc.go
