+++ pkg/s3-proxy/metrics/statusWriter.go
