+++ pkg/s3-proxy/metrics/utils.go
