+++ pkg/s3-proxy/s3client/doc.go
