+++ pkg/s3-proxy/server/middlewares/log-middleware.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/go-chi/chi/middleware"
-	"github.com/oxyno-zeta/s3-proxy/pkg/server/utils"
+	"github.com/oxyno-zeta/s3-proxy/pkg/s3-proxy/server/utils"
 	"github.com/sirupsen/logrus"
 )
