+++ pkg/s3-proxy/server/middlewares/oidc.go
@@ -10,8 +10,8 @@ import (
 
 	oidc "github.com/coreos/go-oidc"
 	"github.com/go-chi/chi"
-	"github.com/oxyno-zeta/s3-proxy/pkg/config"
-	"github.com/oxyno-zeta/s3-proxy/pkg/server/utils"
+	"github.com/oxyno-zeta/s3-proxy/pkg/s3-proxy/config"
+	"github.com/oxyno-zeta/s3-proxy/pkg/s3-proxy/server/utils"
 	"github.com/sirupsen/logrus"
 	"github.com/thoas/go-funk"
