+++ pkg/s3-proxy/version/version.go
