+++ pkg/s3-proxy/version/version_test.go
